requirements.txt
streamlit
openai
python-pptx
pandas
